Deprecate the whole project ;( Telegram has completely rewrote its calls library, now it's 'tgcalls': https://github.com/TelegramMessenger/tgcalls The brand new Python wrapper for it is here: https://github.com/MarshalX/tgcalls
Back in 2019, there already was a Python wrapper called 'pytgvoip', which is mentioned in the ReadMe - that's why I chose 'pytgcalls' as a name for my own implementation of it then. But now Telegram's calls library is named 'tgcalls' as well, and the new wrapper for it is also called 'pytgcalls'. And since GitHub usage statistics is being collected for older repos, it has decided that my repo is the repo of the new wrapper, thus "stealing" statistics from the real author. He kindly asked me to deprecate my project to resolve that problem.

Although I really wanted to finish it someday after finishing the university in this year, now things had changed and the new ready-to-go effort is already here. So, let it be. Life continues.